bench/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "er-bench"
version = "0.1.0"
description = "Extensible Entity Resolution Benchmark Framework"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ICE-ID Team"}
]
dependencies = [
    "numpy>=1.21",
    "pandas>=1.3",
    "scikit-learn>=1.0",
    "pyyaml>=6.0",
    "tqdm>=4.60",
    "joblib>=1.1",
    "requests>=2.25",
]

[project.optional-dependencies]
deep = [
    "torch>=1.10",
    "transformers>=4.20",
    "sentence-transformers>=2.2",
]
all = [
    "xgboost>=1.5",
    "lightgbm>=3.3",
    "dedupe>=2.0",
    "recordlinkage>=0.15",
]

[project.scripts]
er-bench = "bench.cli:main"

[tool.setuptools.packages.find]
where = ["."]

